+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
-## [Unreleased]
+## [1.3.0] - 2020-12-18
 
 ### Added
 
@@ -475,7 +475,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 
 - No changelog available.
 
-[unreleased]: https://github.com/anttikivi/couplet-composer/compare/v1.2.1...HEAD
+[unreleased]: https://github.com/anttikivi/couplet-composer/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/anttikivi/couplet-composer/compare/v1.2.1...v1.3.0
 [1.2.1]: https://github.com/anttikivi/couplet-composer/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/anttikivi/couplet-composer/compare/v1.1.5...v1.2.0
 [1.1.5]: https://github.com/anttikivi/couplet-composer/compare/v1.1.4...v1.1.5